+++ CHANGELOG.md
@@ -7,8 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- title in bold text #27
+- auto hide preview when page too narrow #26
+
 ## [0.2.7] - 2026-03-20
 
+- fix release script when run in jj wc
 - auto scan function for coding agents
 - symbols to show notitification settings globally and pane specific
 - new breadcrum display of location in tmux